+++ src/test/java/org/bigml/binding/PredictionsStepdefs.java
@@ -150,7 +150,7 @@ public class PredictionsStepdefs {
     	JSONObject data = (JSONObject) JSONValue.parse(inputData);  	
     	try {
     		Prediction prediction = context.localModel.predict(
-	    			data, null, (JSONObject) JSONValue.parse(operatingPoint), null, true, null, true);
+	    			data, null, (JSONObject) JSONValue.parse(operatingPoint), null, true, null);
 	    	
 	        context.localModelPrediction = prediction;
     	} catch (Exception e) {
@@ -187,7 +187,7 @@ public class PredictionsStepdefs {
 
     	JSONObject data = (JSONObject) JSONValue.parse(inputData);
     	Prediction prediction = context.localModel.predict(
-    			data, null, null, kind, true, true);
+    			data, null, null, kind, true);
     	
         context.localModelPrediction = prediction;
     }
@@ -259,10 +259,10 @@ public class PredictionsStepdefs {
     	JSONObject data = (JSONObject) JSONValue.parse(inputData);
     	
     	try {
-	    	JSONObject prediction = context.localEnsemble
+    		HashMap<String, Object> prediction = context.localEnsemble
 	                .predict(data, null, null, null,
 	                		(JSONObject) JSONValue.parse(operatingPoint), 
-	                		null, true, true, true);
+	                		null, true, true);
 	    	
 	        context.localPrediction = prediction;
     	} catch (Exception e) {
@@ -298,9 +298,9 @@ public class PredictionsStepdefs {
             throws Exception {
 
     	JSONObject data = (JSONObject) JSONValue.parse(inputData);
-    	JSONObject prediction = context.localEnsemble
+    	HashMap<String, Object> prediction = context.localEnsemble
                 .predict(data, null, null, null,
-                        null, kind, true, true, true);
+                        null, kind, true, true);
     	
         context.localPrediction = prediction;
     }
@@ -360,8 +360,8 @@ public class PredictionsStepdefs {
             throws AuthenticationException {
 
     	JSONObject data = (JSONObject) JSONValue.parse(inputData);
-    	JSONObject prediction = context.localLogisticRegression.predict(
-    			data, null, kind, true, true);
+    	HashMap<String, Object> prediction 
+    		= context.localLogisticRegression.predict(data, null, kind, true);
         context.localPrediction = prediction;
     }
     
@@ -389,8 +389,8 @@ public class PredictionsStepdefs {
     		throws Throwable {
     	
     	JSONObject data = (JSONObject) JSONValue.parse(inputData);
-    	JSONObject prediction = context.localLogisticRegression.predict(
-    			data, null, null, true, true);
+    	HashMap<String, Object> prediction 
+    		= context.localLogisticRegression.predict(data, null, null, true);
     	context.localPrediction = prediction;
     }
     
@@ -451,8 +451,8 @@ public class PredictionsStepdefs {
     	
     	JSONObject data = (JSONObject) JSONValue.parse(inputData);
     	
-    	JSONObject prediction = context.localDeepnet.predict(
-    			data, (JSONObject) JSONValue.parse(operatingPoint), null, true, true);
+    	HashMap<String, Object> prediction = context.localDeepnet.predict(
+    			data, (JSONObject) JSONValue.parse(operatingPoint), null, true);
         context.localPrediction = prediction;
     }
     
@@ -483,8 +483,8 @@ public class PredictionsStepdefs {
             throws AuthenticationException {
 
     	JSONObject data = (JSONObject) JSONValue.parse(inputData);
-    	JSONObject prediction = context.localDeepnet.predict(
-    			data, null, kind, true, true);
+    	HashMap<String, Object> prediction = context.localDeepnet.predict(
+    			data, null, kind, true);
         context.localPrediction = prediction;
     }
     
@@ -511,8 +511,8 @@ public class PredictionsStepdefs {
     		throws Throwable {
     	
     	JSONObject data = (JSONObject) JSONValue.parse(inputData);
-    	JSONObject prediction = context.localDeepnet.predict(
-    			data, null, null, true, true);
+    	HashMap<String, Object> prediction = context.localDeepnet.predict(
+    			data, null, null, true);
     	context.localPrediction = prediction;
     }
     
@@ -588,7 +588,8 @@ public class PredictionsStepdefs {
         JSONObject inputData = (JSONObject) JSONValue.parse(args);
         JSONArray inputDataList = new JSONArray();
         inputDataList.add(inputData);
-        List<MultiVote> votes = context.multiModel.batchPredict(inputDataList, null, true, false, MissingStrategy.LAST_PREDICTION,
+        List<MultiVote> votes = context.multiModel.batchPredict(
+        		inputDataList, null, false, MissingStrategy.LAST_PREDICTION,
                 null, false, true);
 
         Double prediction = (Double) votes.get(0).getPredictions()[0].get("prediction");
@@ -621,7 +622,7 @@ public class PredictionsStepdefs {
     	JSONObject data = (JSONObject) JSONValue.parse(inputData);  	
     	try {
 	        context.localPrediction = context.localFusion.predict(
-	        		data, null, null, true, true);
+	        		data, null, null, true);
     	} catch (Exception e) {
 			e.printStackTrace();
 		}